Fix #981 - Refactor LoRaWAN device controller (#1134)
